@@ -1,17 +1,17 @@
 package me.kahlout.rangebuddy;
 
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
+import com.google.android.material.navigation.NavigationView;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import android.app.Activity;
 import android.view.MenuItem;
 
 import com.google.firebase.analytics.FirebaseAnalytics;
@@ -22,6 +22,8 @@ public class MainActivity extends AppCompatActivity
     Toolbar mToolbar;
     DrawerLayout mDrawer;
     NavigationView mNavigationView;
+
+    public static Activity mActivity;
 
     // Firebase Analytics
     private FirebaseAnalytics mFirebaseAnalytics;
@@ -49,6 +51,25 @@ public class MainActivity extends AppCompatActivity
 
         /// Set first screen
         displaySelectedScreen(R.id.nav_map);
+
+        // Set Activity
+        mActivity = this;
+
+    }
+
+    public static Activity getActivity(){
+
+        return mActivity;
+    }
+
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        if(mActivity==null)
+        {
+            mActivity = MainActivity.this;
+        }
     }
 
     @Override
